src/requirements.txt
flask[async]
flask-cors
firebase-admin
python-dotenv
gunicorn 